fixed opencl int mismatch with size_t

+++ src/layers_cl/batchnorm_layer_cl.h
@@ -38,10 +38,10 @@ class BatchNormLayerCL : public Layer<Dtype>
         int PreCalParams();
 
     private:
-        size_t output_channels;
-        size_t output_width;
-        size_t output_height;
-        size_t channel_block_size;
+        int output_channels;
+        int output_width;
+        int output_height;
+        int channel_block_size;
 
         bool fuse_scale;
         bool fuse_relu;
